// src/serverApi/v3/models/external-tool-configuration-template-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { CustomParameterResponse } from './custom-parameter-response';

/**
 * 
 * @export
 * @interface ExternalToolConfigurationTemplateResponse
 */
export interface ExternalToolConfigurationTemplateResponse {
    /**
     * 
     * @type {string}
     * @memberof ExternalToolConfigurationTemplateResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ExternalToolConfigurationTemplateResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ExternalToolConfigurationTemplateResponse
     */
    logoUrl?: string;
    /**
     * 
     * @type {Array<CustomParameterResponse>}
     * @memberof ExternalToolConfigurationTemplateResponse
     */
    parameters: Array<CustomParameterResponse>;
    /**
     * 
     * @type {number}
     * @memberof ExternalToolConfigurationTemplateResponse
     */
    version: number;
}