// src/serverApi/v3/models/tool-configuration-entry-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HPI Schul-Cloud Server API
 * This is v3 of HPI Schul-Cloud Server. Checkout /docs for v1.
 *
 * The version of the OpenAPI document: 3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ToolConfigurationEntryResponse
 */
export interface ToolConfigurationEntryResponse {
    /**
     * 
     * @type {string}
     * @memberof ToolConfigurationEntryResponse
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ToolConfigurationEntryResponse
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof ToolConfigurationEntryResponse
     */
    logoUrl?: string;
}